Added keyword 'repeat' to test specification
This small change allows tests to be run multiple times without
having to manually create extra tests. It is invoked by adding the
'repeat' value in the 'rtest' declaration of the rxml (it is not required):

...
<rtest name="elastic-random" repeat="10">
...

This will create copies of the tests (in the testresults dir) named

elastic-random_01
elastic-random_02
...
elastic-random_10

and will be run and reported on as individual tests. This will easily allow
random tests to greatly increase the test coverage (and leverage the speed
and small storage footprint of GMD).