runtests: execute name expands environmental vars.
This was necessary because of the cesf tests requiring the same
version of python to run the analyzer as was used to build exowrap.
Now, the test has

  <execute name="$PYTHON ccsf"> {NAME}.exo </execute>

to run the analysis.

I also needed to change the executable wrappers to EXPORT the variable
PYTHON instead of just setting it locally so that test.py can find the
PYTHON variable and expand it.

The 'which()' function was also changed to allow for the exe string of
the form

exe='/path/to/python pyfile.py --opts1 --opts2'

to be given without the function choking. Original functionality has
been preserved.