Removed requirement for 64-bit python
By only removing the block of code in setup.py I was
able to build and run gmd on my little 32-bit netbook.

All tests ran and passed.

On another note, I was able to build it on 32-bit Debian
and only needed to install the standard numpy, scipy,
gfortran, and then libblas-dev and liblapack-dev!