Added a couple analytical benchmarks
Benchmarks:

Added a random test (that is currently set to run 10 times) that
generates a random strain path and calculates the analytical solution
to that.

Added a benchmark based on the paper "A multi-stage return algorithm
for solving the classical damage component of constitutive models for
rocks, ceramics, and other rock-like media", published on 25 September
2009 by Brannon, Leelavanichkul. This is example #2 for perfectly
plastic models. The diff and fail tolerances are much higher than
I would like. Perhaps more work should be put into it.

Also added defaults for the vonmises material model.