Linear Drucker Prager model added
This commit adds the 'pyplastic' model which is just an
isotropic, linear-elastic model with a yield surface
that is parameterized by A1 and A4. It can accurately
mimic 'elastic' and 'vonmises' models as well.

I've added two benchmarks. One is a static benchmark where
it loads in shear until yield and then changes direction to
load in hydrostatic tension until the stress state just
reaches the vertex. The second test is a randomized test
that randomly generates elastic modulii and yield surface
parameters for the LDP model and loads it in such a way
that when it yields the stress state is stationary.