Big new commit, new API, input file format, reorg of code.
COMMENTS
This is a huge change, more than can be fit in to a single commit
message.  The code has been reorganized substantially.  I think it is
much better than before.  Gone are the xml input files.  Now, input
files are python files and models are set up and run more like scripts.
Some of the previous functionality has not been ported, but the new
way is infinitely easier to extend and customize.

TODO:
optimize, permutate, runtests

I'm sure there are others.  When runtests is done and all tests pass,
this will become v2.0