Substantial change in material model installation.
In the old way, each material had an interface file and then some sort
of configuration file.  Now, each material has only the interface file.
it is read at runtime.  I hadn't done it this way in the past because of
speed.  But, I roled my own class browser that is pretty fast.  The
result is that each model provides only the interface file (and any
source files for fortran models).  One limitation that I have imposed is
that each model's interface must be named with one of the following
patterns:

mat_*.py
mat-*.py
*_mat.py
*-mat.py

The regex to find the files is the same that nosetests uses to find test
files - replacing test with mat.

Tests pass, things build, but it's still new.

MUCH CLEANER!