Mimicing is enabled through the rc file
Holy cow! This thing is a work of art! The whole thing came together
with such a pretty, glittery, lace-laden bow that I could not have
hoped for anything better.

If you are curious as to how the magic works, go to core/material.py
around line 355 in the parameter_names() property. When it is
looking for parameter names it is aware of if we are mimicing something
and returns the correct list of names. The model must have a dictionary
of parameter name lists with keys of the models they are associated with.