Random Lin Elas test now much more verbose
I made some changes (that will likely upset you, Tim), but
I regret nothing. The random linear elastic test now writes
the path and the analytical solution to disk. It also has
a logfile for the overall test that has the status of each
realization. In the logfile for each individual realization
it prints the NRMS error for each variable and whether or
not it failed. This makes debugging so much easier.