Plotter uses random colors now.
I'm not sure if this totally breaks what you were going for in
your 140924 commit (8c62036f), but I very much like having my
overlay lines contrast with what they are overlaying.

In theory, it's a good thing when I can't tell the difference
between them, but it is nice to KNOW that there are actually
two different lines plotted, instead of just guessing. Perhaps
there is some way to have colors constant for each variable but
still be able to distinguish them?