Added a demonstration sample input 'hello_world'
There have been several instances where I really wanted something
incredibly simple to show people the beauty and brevity of
defining an mml input. Regrettably, path_cycle.py is just too
big and wordy to really drive home the point (and takes a while
to run).

As with all new programming languages or packages people expect
a 'hello world' example. That's what I've included. It's a nice
simple uniaxial strain problem but is stress-controlled in the
axial direction. Every last superfluous character has been removed
for simplicity. And it runs in about half a second, to boot!

This should help people see the utility and zen-like nature of mml.