Made sure files get closed
I was having problems with 'too many open files' while doing
an optimization run and so went looking everywhere for posible
points where files weren't being closed. In doing so, I made
every last instance of 'open(file)' into the style of 'with .. as ..:"

In the end it didn't help, but I still wanted the changes to
be added to the repo because I think it's more responsible to
make sure that the files get closed.

I also fixed some problems with the mooney-rivlin input.