Fixed a problem with completions
At one point in the file it swapped LAME with SHEAR, only erroring
when you had a negative poissons ratio (and thus a negative LAME
value which would only error when you had a negative poissons ratio
because the shear modulus can't be negative. Do you follow?).

And I did it in this branch because I also made some changes to some
tests so that they would work with the new syntax.

Also, there were problems in elas.py that made it not work when
being called from the command line. Those are now fixed as well.