Added analytical benchmark for DEFGRAD and SYM(L)
The benchmark is a simple isotropic linear elastic benchmark
that systematically goes applies uni-component strain (the three
uniaxial strain then pure shear for each component). It loads and
unloads before starting another component. I made this to benchmark
the computation of the deformation gradient from the input strain
against an analytic solution.

Basically, we've got some serious problems when it comes to
the deformation gradient computation from strain (and, naturally,
problems with the symmetric part of the velocity gradient). If you
want to see, run the benchmark and visualize the output.

I'm sorry for the stupid test name, but I wanted something to set
it apart from all the other linear elastic tests. Also, the test
needs a touch more cleaning up before it's ready to be shelved as
'perfect' and never touched again.