Fixing examples and tensor component code
In femlib/fileio/exodus.py, there's a section where it tries
to parse out tensor components. It was looking for underscores,
I changed it to look for periods. This was caught because one
of the examples didn't run.

Most other examples run. I think break_point.py is the only one
that didn't run for me.

All tests pass.