'mml examples' and optimizer can halt on failure
I added a command line option 'mml examples' which just copies
the examples directory to your current working directory. This
is meant to be used in conjunction with people that
'pip install matmodlab' because no one wants to dig through their
filesystem just to find the examples.

I added an option to the optimizer 'halt_on_err' which is False
as a default. The default does not change the behavior from before.
If set to True, it causes the optimizer to stop on failure
(including KeyboardInterrrupt). I think we should consider making
the default True. Really, if something fails, wouldn't you want
it to halt so you can figure out what failed and handle it gracefully?