An update to Scot's commit of grabbing the examples directory (which I loved).
OVERVIEW:

The examples mml subcommand was replaced by fetch.  The previous
behavior is now

    mml fetch examples

mml fetch does a few other things:

o get copy of the tests

    mml fetch tests

o get copy of the documentation

    mml fetch documentation

o get a copy of a file in the examples directory:

    mml fetch filename

The last command should be updated to not only get the filename, but
also any resources required by the file (for example, an example file
that runs a umat should also get the umat).