Added new method of "installing" a model.
The new method uses the load_material function to load
a material model class in to Matmodlab.  As an example, given a script
that defines a material

class MyMaterial(MaterialModel):
    code

the model can be loaded directly in to Matmodlab by

load_material(MyMaterial)

It is only loaded for the current python session (it is not persistent).

A use case is to develop a model in a notebook.  See
examples/MooneyRivlin.ipynb for an example.