optimizer adds link to final run. fix mps.get()
I added a little thing that, at the end of an optimizaiton, creates a
symlink to the last optimizer run. The symlink is called 'final', so
you can always access the final data using myrun.eval/final/output.dbx
or whatever you want. It is very simply implemented, but not very
satisfying in the way it was done.

Also, when using mps.get() without any args caused a failure. There
was an if-statement that didn't make any sense to me in mmd/simulator.py
around line 360. Look for TJF in a comment to find it (or diff the
commits). My fix seems to work, but it likely didn't fix it in the way
that it should be fixed.