PEP 328-ifying the code base. tabfileio is py3 ready.
PEP 328 deals with absolute/relative imports and tries to help us
be more explicit with our import statements. For the most part, the
changes that I made make it so that we do relative imports throughout
the package instead of re-stating "from matmodlab.mmd.simulator import X"
to just be "from .simulator import X". This also helps take us closer to
being python3 ready.

Speaking of which, tabfileio can now be used with python3. In fact, it is
better to use python3 with it instead of python2 because I had to choose
which one I wanted to support in order to only have one code base. The only
thing that got lost is crazy unicode character support in python2. I
honestly don't think that this is much of a loss at all.

All tests pass using "python setup.py develop" and "python setup.py install"
in a fresh install of anaconda 2.3.