Dynamic stepping - dynamic allocation of Records
I've taken the first step to making dynamic stepping possible. Now, when I
say "dynamic stepping" I'm talking about running the steps when you make them,
not when you do "mps.run()". Today's modification that doesn't change the
functionality is to the Records class. It extends the numpy array for each
step as it is executed. As a test to see how performance is affected, I ran a
simple elastic test with 100 steps of 50 frames each and did it 20 times for
the old way and the new way. This is what I got:

old (avg, stddev): (4.452535999999999, 0.09378110632744752)
new (avg, stddev): (4.438549999999999, 0.10876952008720085)

Performance isn't really affected.