Adding a stress-controlled benchmark
This benchmark randomly generates one leg of an elastic analysis
and uses the analytical solution to run a strain-controlled and then
a stress-controlled simulation. Absolutely everything is random:
elastic modulii, the number of steps, the magnitude of each strain,
everything but the duration (1.0 time unit).

There are some very interesting things in the error values for the
comparisons. Specifically, the error in 'time' is almost always very
nearly 1.0e-8. Why it isn't zero or, at least, very close to 1.0e-14,
I haven't a clue.

The error in the strain is usually evenly distributed along [0.0, 1.0e-8]
and the stresses are normally-distributed around 1.5e-8 with some crazy
outliers every once in a while.